refactor(ipfs): simplify client upload flow around proxy worker
